+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/InsertDepartment.java
@@ -21,9 +21,8 @@ package is.codion.framework.demos.employees.testing.scenarios;
 import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
+import is.codion.framework.domain.entity.test.DefaultEntityFactory;
 import is.codion.swing.framework.model.SwingEntityModel;
-
-import static is.codion.framework.domain.entity.test.EntityTestUtil.createRandomEntity;
 
 // tag::loadTest[]
 public final class InsertDepartment implements Performer<EmployeesAppModel> {
@@ -31,7 +30,7 @@ public final class InsertDepartment implements Performer<EmployeesAppModel> {
 	@Override
 	public void perform(EmployeesAppModel application) throws Exception {
 		SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
-		departmentModel.editModel().set(createRandomEntity(application.entities(), Department.TYPE, null));
+		departmentModel.editModel().set(new DefaultEntityFactory(application.entities()).entity(Department.TYPE));
 		departmentModel.editModel().insert();
 	}
 }