+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/SelectDepartment.java
@@ -18,14 +18,11 @@
  */
 package is.codion.framework.demos.employees.testing.scenarios;
 
-import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
 
-import static is.codion.swing.framework.model.tools.loadtest.EntityLoadTestUtil.selectRandomRow;
-
 // tag::loadTest[]
-public final class SelectDepartment implements Performer<EmployeesAppModel> {
+public final class SelectDepartment extends AbstractPerformer {
 
 	@Override
 	public void perform(EmployeesAppModel application) {