+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/UpdateEmployee.java
@@ -18,21 +18,19 @@
  */
 package is.codion.framework.demos.employees.testing.scenarios;
 
-import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.domain.Employees.Employee;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
 import is.codion.framework.domain.entity.Entity;
+import is.codion.framework.domain.entity.test.DefaultEntityFactory;
+import is.codion.framework.domain.entity.test.EntityTestUnit.EntityFactory;
 import is.codion.swing.framework.model.SwingEntityModel;
 
 import java.util.Random;
 
-import static is.codion.framework.domain.entity.test.EntityTestUtil.randomize;
-import static is.codion.swing.framework.model.tools.loadtest.EntityLoadTestUtil.selectRandomRow;
-
 // tag::loadTest[]
-public final class UpdateEmployee implements Performer<EmployeesAppModel> {
+public final class UpdateEmployee extends AbstractPerformer {
 
 	private final Random random = new Random();
 
@@ -41,18 +39,19 @@ public final class UpdateEmployee implements Performer<EmployeesAppModel> {
 		SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
 		selectRandomRow(departmentModel.tableModel());
 		SwingEntityModel employeeModel = departmentModel.detailModel(Employee.TYPE);
+		EntityFactory entityFactory = new DefaultEntityFactory(application.entities());
 		if (employeeModel.tableModel().getRowCount() > 0) {
 			EntityConnection connection = employeeModel.connection();
 			connection.startTransaction();
 			try {
 				selectRandomRow(employeeModel.tableModel());
 				Entity selected = employeeModel.tableModel().selectionModel().getSelectedItem();
-				randomize(application.entities(), selected, null);
+				entityFactory.modify(selected);
 				employeeModel.editModel().set(selected);
 				employeeModel.editModel().update();
 				selectRandomRow(employeeModel.tableModel());
 				selected = employeeModel.tableModel().selectionModel().getSelectedItem();
-				randomize(application.entities(), selected, null);
+				entityFactory.modify(selected);
 				employeeModel.editModel().set(selected);
 				employeeModel.editModel().update();
 			}