+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/domain/StoreTest.java
@@ -18,13 +18,13 @@
  */
 package is.codion.framework.demos.manual.store.domain;
 
-import is.codion.common.db.exception.DatabaseException;
 import is.codion.framework.demos.manual.store.domain.Store.Address;
 import is.codion.framework.demos.manual.store.domain.Store.Customer;
 import is.codion.framework.demos.manual.store.domain.Store.CustomerAddress;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
+import is.codion.framework.domain.entity.test.DefaultEntityFactory;
 import is.codion.framework.domain.entity.test.EntityTestUnit;
 
 import org.junit.jupiter.api.Test;
@@ -35,8 +35,10 @@ import java.util.UUID;
 // tag::storeTest[]
 public class StoreTest extends EntityTestUnit {
 
+	private static final Store DOMAIN = new Store();
+
 	public StoreTest() {
-		super(new Store());
+		super(DOMAIN, new StoreEntityFactory());
 	}
 
 	@Test
@@ -54,65 +56,71 @@ public class StoreTest extends EntityTestUnit {
 		test(CustomerAddress.TYPE);
 	}
 
-	@Override
-	protected Entity initializeForeignKeyEntity(ForeignKey foreignKey,
-																							Map<ForeignKey, Entity> foreignKeyEntities)
-					throws DatabaseException {
-		//see if the currently running test requires an ADDRESS entity
-		if (foreignKey.referencedType().equals(Address.TYPE)) {
-			return entities().builder(Address.TYPE)
-							.with(Address.ID, 21L)
-							.with(Address.STREET, "One Way")
-							.with(Address.CITY, "Sin City")
-							.build();
+	private static final class StoreEntityFactory extends DefaultEntityFactory {
+
+		private StoreEntityFactory() {
+			super(DOMAIN.entities());
 		}
 
-		return super.initializeForeignKeyEntity(foreignKey, foreignKeyEntities);
-	}
+		@Override
+		public Entity foreignKeyEntity(ForeignKey foreignKey,
+																	 Map<ForeignKey, Entity> foreignKeyEntities) {
+			//see if the currently running test requires an ADDRESS entity
+			if (foreignKey.referencedType().equals(Address.TYPE)) {
+				return entities().builder(Address.TYPE)
+								.with(Address.ID, 21L)
+								.with(Address.STREET, "One Way")
+								.with(Address.CITY, "Sin City")
+								.build();
+			}
 
-	@Override
-	protected Entity initializeTestEntity(EntityType entityType,
-																				Map<ForeignKey, Entity> foreignKeyEntities) {
-		if (entityType.equals(Address.TYPE)) {
-			//Initialize an entity representing the table STORE.ADDRESS,
-			//which can be used for the testing
-			return entities().builder(Address.TYPE)
-							.with(Address.ID, 42L)
-							.with(Address.STREET, "Street")
-							.with(Address.CITY, "City")
-							.build();
-		}
-		else if (entityType.equals(Customer.TYPE)) {
-			//Initialize an entity representing the table STORE.CUSTOMER,
-			//which can be used for the testing
-			return entities().builder(Customer.TYPE)
-							.with(Customer.ID, UUID.randomUUID().toString())
-							.with(Customer.FIRST_NAME, "Robert")
-							.with(Customer.LAST_NAME, "Ford")
-							.with(Customer.ACTIVE, true)
-							.build();
-		}
-		else if (entityType.equals(CustomerAddress.TYPE)) {
-			return entities().builder(CustomerAddress.TYPE)
-							.with(CustomerAddress.CUSTOMER_FK, foreignKeyEntities.get(Customer.TYPE))
-							.with(CustomerAddress.ADDRESS_FK, foreignKeyEntities.get(Address.TYPE))
-							.build();
+			return super.foreignKeyEntity(foreignKey, foreignKeyEntities);
 		}
 
-		return super.initializeTestEntity(entityType, foreignKeyEntities);
-	}
+		@Override
+		public Entity entity(EntityType entityType,
+												 Map<ForeignKey, Entity> foreignKeyEntities) {
+			if (entityType.equals(Address.TYPE)) {
+				//Initialize an entity representing the table STORE.ADDRESS,
+				//which can be used for the testing
+				return entities().builder(Address.TYPE)
+								.with(Address.ID, 42L)
+								.with(Address.STREET, "Street")
+								.with(Address.CITY, "City")
+								.build();
+			}
+			else if (entityType.equals(Customer.TYPE)) {
+				//Initialize an entity representing the table STORE.CUSTOMER,
+				//which can be used for the testing
+				return entities().builder(Customer.TYPE)
+								.with(Customer.ID, UUID.randomUUID().toString())
+								.with(Customer.FIRST_NAME, "Robert")
+								.with(Customer.LAST_NAME, "Ford")
+								.with(Customer.ACTIVE, true)
+								.build();
+			}
+			else if (entityType.equals(CustomerAddress.TYPE)) {
+				return entities().builder(CustomerAddress.TYPE)
+								.with(CustomerAddress.CUSTOMER_FK, foreignKeyEntities.get(Customer.TYPE))
+								.with(CustomerAddress.ADDRESS_FK, foreignKeyEntities.get(Address.TYPE))
+								.build();
+			}
 
-	@Override
-	protected void modifyEntity(Entity testEntity, Map<ForeignKey, Entity> foreignKeyEntities) {
-		if (testEntity.entityType().equals(Address.TYPE)) {
-			testEntity.put(Address.STREET, "New Street");
-			testEntity.put(Address.CITY, "New City");
+			return super.entity(entityType, foreignKeyEntities);
 		}
-		else if (testEntity.entityType().equals(Customer.TYPE)) {
-			//It is sufficient to change the value of a single property, but the more, the merrier
-			testEntity.put(Customer.FIRST_NAME, "Jesse");
-			testEntity.put(Customer.LAST_NAME, "James");
-			testEntity.put(Customer.ACTIVE, false);
+
+		@Override
+		public void modify(Entity entity, Map<ForeignKey, Entity> foreignKeyEntities) {
+			if (entity.entityType().equals(Address.TYPE)) {
+				entity.put(Address.STREET, "New Street");
+				entity.put(Address.CITY, "New City");
+			}
+			else if (entity.entityType().equals(Customer.TYPE)) {
+				//It is sufficient to change the value of a single property, but the more, the merrier
+				entity.put(Customer.FIRST_NAME, "Jesse");
+				entity.put(Customer.LAST_NAME, "James");
+				entity.put(Customer.ACTIVE, false);
+			}
 		}
 	}
 }