+++ framework/domain-test/src/main/java/is/codion/framework/domain/entity/test/DefaultEntityFactory.java
@@ -18,13 +18,16 @@
  */
 package is.codion.framework.domain.entity.test;
 
+import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.item.Item;
+import is.codion.framework.db.EntityConnection;
 import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.AttributeDefinition;
+import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.attribute.ColumnDefinition;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.framework.domain.entity.attribute.ForeignKeyDefinition;
@@ -38,23 +41,27 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.OffsetDateTime;
 import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Random;
 import java.util.function.Function;
 import java.util.stream.IntStream;
 
+import static java.util.Collections.emptyMap;
+import static java.util.Collections.singletonList;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toList;
 
 /**
- * Utility methods for creating and manipulating Entity instances for testing purposes.
+ * Handles creating and manipulating Entity instances for testing purposes.
  */
-public final class EntityTestUtil {
+public class DefaultEntityFactory implements EntityTestUnit.EntityFactory {
 
-	private static final Logger LOG = LoggerFactory.getLogger(EntityTestUtil.class);
+	private static final Logger LOG = LoggerFactory.getLogger(DefaultEntityFactory.class);
 
 	private static final int MININUM_RANDOM_NUMBER = -10_000;
 	private static final int MAXIMUM_RANDOM_NUMBER = 10_000;
@@ -62,7 +69,75 @@ public final class EntityTestUtil {
 	private static final String ALPHA_NUMERIC = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 	private static final Random RANDOM = new Random();
 
-	private EntityTestUtil() {}
+	private final Entities entities;
+
+	public DefaultEntityFactory(Entities entities) {
+		this.entities = requireNonNull(entities);
+	}
+
+	@Override
+	public Entity entity(EntityType entityType) {
+		return entity(entityType, emptyMap());
+	}
+
+	@Override
+	public Entity entity(EntityType entityType, Map<ForeignKey, Entity> foreignKeyEntities) {
+		return randomEntity(entityType, foreignKeyEntities);
+	}
+
+	@Override
+	public Entity foreignKeyEntity(ForeignKey foreignKey, Map<ForeignKey, Entity> foreignKeyEntities) {
+		if (entities.definition(requireNonNull(foreignKey).referencedType()).readOnly()) {
+			return null;
+		}
+
+		return randomEntity(foreignKey.referencedType(), foreignKeyEntities);
+	}
+
+	@Override
+	public void modify(Entity entity) {
+		modify(entity, null);
+	}
+
+	@Override
+	public void modify(Entity entity, Map<ForeignKey, Entity> foreignKeyEntities) {
+		randomize(entity, foreignKeyEntities);
+	}
+
+	@Override
+	public Map<ForeignKey, Entity> foreignKeyEntities(EntityType entityType,
+																										Map<ForeignKey, Entity> foreignKeyEntities,
+																										EntityConnection connection) throws DatabaseException {
+		List<ForeignKey> foreignKeys = new ArrayList<>(entities.definition(entityType).foreignKeys().get());
+		//we have to start with non-self-referential ones
+		foreignKeys.sort((fk1, fk2) -> !fk1.referencedType().equals(entityType) ? -1 : 1);
+		for (ForeignKey foreignKey : foreignKeys) {
+			EntityType referencedEntityType = foreignKey.referencedType();
+			if (!foreignKeyEntities.containsKey(foreignKey)) {
+				if (!Objects.equals(entityType, referencedEntityType)) {
+					foreignKeyEntities.put(foreignKey, null);//short circuit recursion, value replaced below
+					foreignKeyEntities(referencedEntityType, foreignKeyEntities, connection);
+				}
+				Entity referencedEntity = foreignKeyEntity(foreignKey, foreignKeyEntities);
+				if (referencedEntity != null) {
+					foreignKeyEntities.put(foreignKey, insertOrSelect(referencedEntity, connection));
+				}
+			}
+		}
+
+		return foreignKeyEntities;
+	}
+
+	/**
+	 * @return the underlying {@link Entities} instance
+	 */
+	protected final Entities entities() {
+		return entities;
+	}
+
+	private Entity randomEntity(EntityType entityType, Map<ForeignKey, Entity> foreignKeyEntities) {
+		return createRandomEntity(entityType, foreignKeyEntities);
+	}
 
 	/**
 	 * @param entities the domain model entities
@@ -70,9 +145,8 @@ public final class EntityTestUtil {
 	 * @param referenceEntities entities referenced by the given foreign key
 	 * @return an Entity instance containing randomized values, based on the attribute definitions
 	 */
-	public static Entity createRandomEntity(Entities entities, EntityType entityType,
-																					Map<ForeignKey, Entity> referenceEntities) {
-		return createEntity(entities, entityType, definition -> createRandomValue(definition, referenceEntities));
+	private Entity createRandomEntity(EntityType entityType, Map<ForeignKey, Entity> referenceEntities) {
+		return createEntity(entityType, definition -> createValue(definition, referenceEntities));
 	}
 
 	/**
@@ -81,11 +155,10 @@ public final class EntityTestUtil {
 	 * @param valueProvider the value provider
 	 * @return an Entity instance with insertable attributes populated with values provided by the given value provider
 	 */
-	public static Entity createEntity(Entities entities, EntityType entityType, Function<AttributeDefinition<?>, Object> valueProvider) {
-		requireNonNull(entities);
+	private Entity createEntity(EntityType entityType, Function<Attribute<?>, Object> valueProvider) {
 		requireNonNull(entityType);
 		Entity entity = entities.entity(entityType);
-		populateEntity(entity, insertableColumnDefinitions(entities.definition(entityType)), valueProvider);
+		populateEntity(entity, insertableColumns(entities.definition(entityType)), valueProvider);
 
 		return entity;
 	}
@@ -97,23 +170,23 @@ public final class EntityTestUtil {
 	 * @param entity the entity to randomize
 	 * @param foreignKeyEntities the entities referenced via foreign keys
 	 */
-	public static void randomize(Entities entities, Entity entity, Map<ForeignKey, Entity> foreignKeyEntities) {
-		requireNonNull(entities);
+	private void randomize(Entity entity, Map<ForeignKey, Entity> foreignKeyEntities) {
 		requireNonNull(entity);
 		populateEntity(entity,
-						updatableColumnDefinitions(entity.definition()),
-						attributeDefinition -> createRandomValue(attributeDefinition, foreignKeyEntities));
+						updatableColumns(entity.definition()),
+						attributeDefinition -> createValue(attributeDefinition, foreignKeyEntities));
 	}
 
 	/**
 	 * Creates a random value for the given attribute.
-	 * @param attributeDefinition the attribute definition
+	 * @param attribute the attribute
 	 * @param referenceEntities entities referenced by the given attribute
 	 * @param <T> the attribute value type
 	 * @return a random value
 	 */
-	public static <T> T createRandomValue(AttributeDefinition<T> attributeDefinition, Map<ForeignKey, Entity> referenceEntities) {
-		requireNonNull(attributeDefinition, "attributeDefinition");
+	protected <T> T createValue(Attribute<T> attribute, Map<ForeignKey, Entity> referenceEntities) {
+		requireNonNull(attribute, "attribute");
+		AttributeDefinition<T> attributeDefinition = entities.definition(attribute.entityType()).attributes().definition(attribute);
 		try {
 			if (attributeDefinition instanceof ForeignKeyDefinition) {
 				return (T) referenceEntity(((ForeignKeyDefinition) attributeDefinition).attribute(), referenceEntities);
@@ -121,7 +194,6 @@ public final class EntityTestUtil {
 			if (!attributeDefinition.items().isEmpty()) {
 				return randomItem(attributeDefinition);
 			}
-			Attribute<?> attribute = attributeDefinition.attribute();
 			if (attribute.type().isBoolean()) {
 				return (T) Boolean.valueOf(RANDOM.nextBoolean());
 			}
@@ -173,32 +245,34 @@ public final class EntityTestUtil {
 		}
 	}
 
-	private static void populateEntity(Entity entity, Collection<ColumnDefinition<?>> columnDefinitions,
-																		 Function<AttributeDefinition<?>, Object> valueProvider) {
+	private static void populateEntity(Entity entity, Collection<Column<?>> columns,
+																		 Function<Attribute<?>, Object> valueProvider) {
 		requireNonNull(valueProvider, "valueProvider");
 		EntityDefinition definition = entity.definition();
-		for (ColumnDefinition<?> columnDefinition : columnDefinitions) {
-			if (!definition.foreignKeys().foreignKeyColumn(columnDefinition.attribute())) {
-				entity.put((Attribute<Object>) columnDefinition.attribute(), valueProvider.apply(columnDefinition));
+		for (Column<?> column : columns) {
+			if (!definition.foreignKeys().foreignKeyColumn(column)) {
+				entity.put((Attribute<Object>) column, valueProvider.apply(column));
 			}
 		}
-		for (ForeignKeyDefinition foreignKeyDefinition : entity.definition().foreignKeys().definitions()) {
-			Entity value = (Entity) valueProvider.apply(foreignKeyDefinition);
+		for (ForeignKey foreignKey : entity.definition().foreignKeys().get()) {
+			Entity value = (Entity) valueProvider.apply(foreignKey);
 			if (value != null) {
-				entity.put(foreignKeyDefinition.attribute(), value);
+				entity.put(foreignKey, value);
 			}
 		}
 	}
 
-	private static List<ColumnDefinition<?>> insertableColumnDefinitions(EntityDefinition entityDefinition) {
+	private static List<Column<?>> insertableColumns(EntityDefinition entityDefinition) {
 		return entityDefinition.columns().definitions().stream()
 						.filter(column -> column.insertable() && (!entityDefinition.primaryKey().generated() || !column.primaryKey()))
+						.map(ColumnDefinition::attribute)
 						.collect(toList());
 	}
 
-	private static List<ColumnDefinition<?>> updatableColumnDefinitions(EntityDefinition entityDefinition) {
+	private static List<Column<?>> updatableColumns(EntityDefinition entityDefinition) {
 		return entityDefinition.columns().definitions().stream()
 						.filter(column -> column.updatable() && !column.primaryKey())
+						.map(ColumnDefinition::attribute)
 						.collect(toList());
 	}
 
@@ -278,5 +352,29 @@ public final class EntityTestUtil {
 						MAXIMUM_RANDOM_NUMBER : Math.min(attributeDefinition.maximumValue().doubleValue(), MAXIMUM_RANDOM_NUMBER);
 
 		return RANDOM.nextDouble() * (max - min) + min;
+	}
+
+	/**
+	 * Inserts or selects the given entity if it exists and returns the result
+	 * @param entity the entity to initialize
+	 * @param connection the connection to use
+	 * @return the entity
+	 * @throws DatabaseException in case of an exception
+	 */
+	private static Entity insertOrSelect(Entity entity, EntityConnection connection) throws DatabaseException {
+		try {
+			if (entity.primaryKey().isNotNull()) {
+				Collection<Entity> selected = connection.select(singletonList(entity.primaryKey()));
+				if (!selected.isEmpty()) {
+					return selected.iterator().next();
+				}
+			}
+
+			return connection.insertSelect(entity);
+		}
+		catch (DatabaseException e) {
+			LOG.error("EntityTestUnit.insertOrSelect()", e);
+			throw new DatabaseException(e.getMessage() + ": " + entity);
+		}
 	}
 }